+++ dwds/lib/dwds.dart
@@ -4,17 +4,17 @@
 
 import 'dart:async';
 
-import 'package:dwds/data/build_result.dart';
-import 'package:dwds/src/utilities/shared.dart';
 import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
 import 'package:shelf/shelf.dart';
 import 'package:webkit_inspection_protocol/webkit_inspection_protocol.dart';
 
+import 'data/build_result.dart';
 import 'src/connections/app_connection.dart';
 import 'src/connections/debug_connection.dart';
 import 'src/handlers/dev_handler.dart';
 import 'src/handlers/injected_handler.dart';
+import 'src/loaders/strategy.dart';
 import 'src/readers/asset_reader.dart';
 import 'src/servers/devtools.dart';
 import 'src/servers/extension_backend.dart';
@@ -22,6 +22,9 @@ import 'src/servers/extension_backend.dart';
 export 'src/connections/app_connection.dart' show AppConnection;
 export 'src/connections/debug_connection.dart' show DebugConnection;
 export 'src/handlers/dev_handler.dart' show AppConnectionException;
+export 'src/loaders/legacy.dart' show LegacyStrategy;
+export 'src/loaders/require.dart' show RequireStrategy;
+export 'src/loaders/strategy.dart' show LoadStrategy, ReloadConfiguration;
 export 'src/readers/asset_reader.dart' show AssetReader;
 export 'src/readers/frontend_server_asset_reader.dart'
     show FrontendServerAssetReader;
@@ -31,8 +34,6 @@ export 'src/services/chrome_proxy_service.dart' show ChromeDebugException;
 typedef LogWriter = void Function(Level, String);
 typedef ConnectionProvider = Future<ChromeConnection> Function();
 typedef UrlEncoder = Future<String> Function(String url);
-enum ReloadConfiguration { none, hotReload, hotRestart, liveReload }
-enum ModuleStrategy { requireJS, legacy }
 
 /// The Dart Web Debug Service.
 class Dwds {
@@ -70,27 +71,25 @@ class Dwds {
     @required AssetReader assetReader,
     @required Stream<BuildResult> buildResults,
     @required ConnectionProvider chromeConnection,
+    @required LoadStrategy loadStrategy,
     @required bool enableDebugging,
+    bool enableDebugExtension,
     String hostname,
-    ReloadConfiguration reloadConfiguration,
     bool useSseForDebugProxy,
     bool serveDevTools,
     LogWriter logWriter,
     bool verbose,
-    bool enableDebugExtension,
-    ModuleStrategy moduleStrategy,
     UrlEncoder urlEncoder,
     @deprecated bool restoreBreakpoints,
   }) async {
     hostname ??= 'localhost';
-    reloadConfiguration ??= ReloadConfiguration.none;
     enableDebugging ??= true;
     enableDebugExtension ??= false;
     useSseForDebugProxy ??= true;
     serveDevTools ??= true;
     logWriter ??= (level, message) => print(message);
     verbose ??= false;
-    globalModuleStrategy = moduleStrategy ?? ModuleStrategy.requireJS;
+    globalLoadStrategy = loadStrategy;
     restoreBreakpoints ??= false;
 
     DevTools devTools;
@@ -130,7 +129,6 @@ class Dwds {
 
     return Dwds._(
       createInjectedHandler(
-        reloadConfiguration,
         extensionUri: extensionUri,
         urlEncoder: urlEncoder,
       ),